Add gitauth,github,bitauth,and bitbucket modules
The scaffold module is currently handling a lot of tasks including version control systems authentication. Implementing vendor specific functionalities make the scaffold module barely mantainable, furthermore, this complicates the code. To combat this and stick to Anzii's phillosophy, we've decided to refactor the scaffold module to implement minimal functionality that is in line with its intended text.

Afther this commit, github and bitbucket functionalities will be handled by separate modules.